add parallel file download

+++ src/com/machina/HttpResult.kt
@@ -4,7 +4,7 @@ data class HttpResult(
     val code: Int,
     val status: String,
     val contentType: String = "",
-    val contentLength: Int = 0,
+    val contentLength: Long = 0,
     val contentHeader: Any = "",
     val content: Any = "",
     val basicAuth: Int
